+++ app/services/data_collector.py
